+++ internal/webhook/doc.go
@@ -1,0 +1,2 @@
+// Todo in endpoints
+package webhook
