+++ pkg/repository/postgres/logs.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import "time"
